지훈 - 별점 평균 관련 이슈 수정

+++ watflix/src/main/java/org/kosta/watflix/model/service/ContentsService.java
@@ -26,5 +26,5 @@ public interface ContentsService {
 	void sContentsLikeRevmoew(String reviewNo);
 	//컨텐츠추천여부검사를위한 서비스
 	//각 content에 contents_avg_stars 넣기
-	void sUpdateAvgStar(float avgStars, String contentsNo);
+	void sUpdateAvgStar(double avgStars, String contentsNo);
 }
